+++ components/Modal/Modal.module.scss
@@ -1,15 +1,14 @@
 .overlay {
   position: fixed;
-  inset: 0;
   display: flex;
+  inset: 0;
   align-items: center;
   justify-content: center;
   background: rgba(0, 0, 0, 0.5);
-  
 }
 
 .modal {
-  position: fixed;
+  position: absolute;
   width: 90%;
   top: 0;
   bottom: 0;
@@ -22,15 +21,14 @@
 }
 
 .body {
+  position: relative;
   background-color: var(--BDS-Grayscale-1);
   padding: 2rem 2.4rem;
-  padding-bottom: 9.2rem;
-}
-
-// 960px 까지 화면 width의 90%가 모달폭
-@media (min-width: 768) {
 }
 
 // 961px 부터 화면 width의 70%가 모달폭
-@media (min-width: 961) {
+@media (min-width: 961px) {
+  .modal {
+    width: 70%;
+  }
 }
