+++ stores/useModalStore.ts
@@ -1,18 +1,21 @@
 import { create } from 'zustand';
 
-// 1차 레이어?? 모달 생길 시 추가
-export type ModalName = '' | 'write' | 'read';
+export type ModalName = 'write' | 'read';
 
 interface ModalStore {
+  postId: number;
   isOpen: boolean;
   name: string;
+  setPostId: (postId: number) => void;
   toggleModal: () => void;
   setCurrentName: (modalName: ModalName) => void;
 }
 
 const useModalStore = create<ModalStore>((set) => ({
+  postId: 0,
   isOpen: false,
   name: '',
+  setPostId: (postId) => set(() => ({ postId })),
   toggleModal: () => set((state) => ({ isOpen: !state.isOpen })),
   setCurrentName: (modalName) => set(() => ({ name: modalName })),
 }));
