+++ styles/global-styles/font.css
@@ -1,9 +1,6 @@
-@import url('/fonts/pretendard/Pretendard.css');
-@import url('/fonts/montserrat/Montserrat.css');
-@import url('/fonts/inter/inter.css');
 /* 로컬에서 어떻게 가져오는지 모르겠삼 */
-@import url('https://fonts.googleapis.com/css2?family=IBM+Plex+Sans+KR&display=swap');
+@import url(../../public/fonts/ibm-plex-sans-kr/IBM_Plex_Sans_KR.css);
 
 * {
-  font-family: 'IBM Plex Sans KR', Pretendard, Montserrat, inter, sans-serif;
+  font-family: "IBM Plex Sans KR";
 }
