+++ styles/global-styles/global.css
@@ -2,7 +2,6 @@
 @import url('font.css');
 @import url('theme.css');
 
-
 :root {
   font-size: 62.5%;
   /* 1rem = 10px */
